Revert "Deploy localization."
