--- assets/icons/manifest.json
@@ -1,12 +1,0 @@
-{
-	"name": "OKCandidate",
-	"icons": [
-		{
-			"src": "\/icons\/android-chrome-192x192.png",
-			"sizes": "192x192",
-			"type": "image\/png"
-		}
-	],
-	"theme_color": "#ffffff",
-	"display": "standalone"
-}
