+++ assets/img/candidates/mcclellan.jpg
