+++ assets/img/candidates/stewart.jpg
