+++ assets/img/candidates/turner.jpg
