+++ assets/js/components/ecosystems/WardFinderPage.jsx
@@ -5,15 +5,13 @@ import {
 } from 'react-redux'
 
 import {
-  toggleWardfinderWardDropdown,
-  selectGeography
+  toggleWardfinderWardDropdown
 } from './../../redux/survey/survey-actions'
 
 import {
   Col,
   Grid,
-  Row,
-  Button
+  Row
 } from 'react-bootstrap'
 
 import Card from './../atoms/Card.jsx'
@@ -75,17 +73,10 @@ class WardFinderPage extends Component {
 
   constructor(props) {
     super(props)
-    this.state = {
-      geolocation: false
-    }
   }
 
   toggleWardDropdown() {
     this.props.dispatch(toggleWardfinderWardDropdown())
-  }
-
-  skipGeolocation() {
-    this.props.dispatch(selectGeography(1))
   }
 
   render() {
@@ -107,44 +98,23 @@ class WardFinderPage extends Component {
             <Row>
               <Col xs={12} sm={6} smOffset={3}>
                 <Card style={style.card}>
-
+                  <h2>Get Started!</h2>
+                  <p>
+                    We can find candidates running for office in your area by using any Norfolk
+                    street address.
+                  </p>
+                  <WardFinderAddress
+                    ward={this.props.ward}
+                    dispatch={this.props.dispatch} />
+                  <a onClick={this.toggleWardDropdown.bind(this)}>Other options</a>
                   {
-                    this.state.geolocation ?
-                    <p>
-                      We can find candidates running for office in your area by using any Virginia Beach
-                      street address.
-                    </p>
+                    this.props.ward.showWardFinderDropdown ?
+                    <WardFinderDropdown
+                      ward={this.props.ward}
+                      dispatch={this.props.dispatch} />
                     :
-                    <p>
-                      Click to start answering questions and find your perfect match!
-                    </p>
+                    null
                   }
-
-                  {
-                    this.state.geolocation ?
-                      <div>
-                        <WardFinderAddress
-                          ward={this.props.ward}
-                          dispatch={this.props.dispatch} />
-                        <a onClick={this.toggleWardDropdown.bind(this)}>Other options</a>
-                        {
-                          this.props.ward.showWardFinderDropdown ?
-                          <WardFinderDropdown
-                            ward={this.props.ward}
-                            dispatch={this.props.dispatch} />
-                          :
-                          null
-                        }
-                      </div>
-                    :
-                      <div>
-                        <Button
-                          onClick={this.skipGeolocation.bind(this)}
-                          bsStyle="primary"
-                          bsSize="large">Get Started</Button>
-                      </div>
-                  }
-
                 </Card>
               </Col>
             </Row>
