+++ assets/js/components/organisms/SurveyCompletionIndicator.jsx
@@ -6,7 +6,6 @@ import {
 } from 'react-bootstrap'
 
 import colors from './../style/colors'
-import { getMatchLevel } from './../utils/matchLevel'
 import Card from './../atoms/Card.jsx'
 
 const style = {
@@ -52,7 +51,44 @@ import SurveyCompletionIndicatorHeart from './../atoms/SurveyCompletionIndicator
 class SurveyCompletionIndicator extends Component {
 
   getMatchLevel(answered, total) {
-    return getMatchLevel(answered, total)
+
+    const third = total / 3 % 1 === 0 ? total / 3 : Math.round(total / 3)
+
+    const result = {
+      level: 0,
+      mood: ':(',
+      text: `Answer ${third - answered % third} questions for an OK rating`,
+      progress: Math.round(answered % third / third * 100)
+    }
+
+    if (answered >= third && answered < (2 * third)) {
+      return Object.assign({}, result, {
+        level: 1,
+        mood: 'OK!',
+        text: `Answer ${third - answered % third} questions for a good rating`
+      })
+    }
+
+    else if (answered >= (2 * third) && answered < total) {
+      return Object.assign({}, result, {
+        level: 2,
+        mood: 'Good!',
+        text: `Answer ${total - answered} questions for the best rating`
+      })
+    }
+
+    else if (answered === total) {
+      return Object.assign({}, result, {
+        level: 3,
+        mood: 'Best!',
+        text: 'Good to go!',
+        progress: 100
+      })
+    }
+
+    else {
+      return result
+    }
   }
 
   render() {
