+++ assets/style/main.css
@@ -1,7 +1,7 @@
 /* General Body Styles */
 body, html {
   background-color: #D8E9F0;
-  background-image: url("/img/pattern-60-lines.png");
+  background-image: url("http://www.transparenttextures.com/patterns/60-lines.png");
 }
 
 /* Survey Question Styles */
