--- ci-frontend.test.yml
@@ -1,27 +1,0 @@
-sut:
-  build: ./
-  command: npm run test-frontend
-  environment:
-    - OKC_DB_HOST=okcpostgres
-    - OKC_DB_NAME=okc
-    - OKC_DB_USER=okc
-    - OKC_DB_PASSWORD=complicatedPassword
-    - SMARTY_STREETS_AUTH_ID=GUID-FROM-SMARTY-STREETS
-    - SMARTY_STREETS_AUTH_TOKEN=tokenIdFromSmartyStreets
-    - NODE_ENV=development
-  ports:
-    - '8000:8000'
-  links:
-    - okcpostgres
-okcpostgres:
-  image: sameersbn/postgresql:latest
-  environment:
-    - PG_PASSWORD=rootPSQLPasswordYouWant
-    - DB_NAME=okc
-    - DB_PASS=complicatedPassword
-    - DB_USER=okc
-  volumes:
-    - postgres-data:/var/lib/postgresql
-  ports:
-    - '5432:5432'
-  mem_limit: 500M
